Auth Working Journal

+++ app/Http/Controllers/PostsController.php
@@ -4,9 +4,15 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Post;
+use App\User;
 
 class PostsController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'About']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -48,6 +54,7 @@ class PostsController extends Controller
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
         $post->save();
 
         return redirect('/posts')->with('success', 'Journal Created');
@@ -74,6 +81,9 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post =  Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -110,8 +120,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
         $post->delete();
-        return redirect('/posts')->with('success', 'Journal Entry Removed');
+        return redirect('/posts')->with('success', 'Journal Entry');
 
     }
 }
